requirements.txt
appdirs==1.4.4
certifi==2020.6.20
decorator==4.4.2
easydict==1.9
future==0.18.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.0
onnx==1.5.0
onnxruntime==1.2.0
opencv-python==4.3.0.36
Pillow==7.2.0
protobuf==3.12.2
pycuda==2019.1.2
pytools==2020.3.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
torch==1.5.1
torchvision==0.6.1
typing==3.7.4.1
typing-extensions==3.7.4.2

